backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.5.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
marshmallow==3.19.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0.post2
requests==2.28.2
pytest==7.3.1
pytest-cov==4.1.0
celery==5.2.7
redis==4.5.4
sendgrid==6.10.0
dadata==21.10.1
werkzeug==2.3.7
alembic==1.11.3
pyjwt==2.6.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0